Moved filterArgs to utility methods, added default value to aws region for s3 module, added documentation to the filterArgs function
