+++ libs/shared/decorators/src/lib/log-and-handle.decorator.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Transaction } from 'sequelize'
-import { handleException } from '../lib/utils'
+import { filterArgs, handleException } from '../lib/utils'
 import { logger } from '@dmr.is/logging'
 
 type LogAndHandleOptions = {
@@ -23,30 +23,7 @@ export function LogAndHandle(
     const originalMethod = descriptor.value
 
     descriptor.value = async function (...args: any[]) {
-      const filteredArgs = args.filter((arg) => {
-        const isTransaction = arg instanceof Transaction
-        const isBuffer = Buffer.isBuffer(arg.buffer) // filter out arguments with buffer / files
-
-        if (Array.isArray(arg)) {
-          const isTransactionOrBuffer = arg.filter((a) => {
-            const isTransaction = a instanceof Transaction
-            const isBuffer = Buffer.isBuffer(a.buffer) // filter out arguments with buffer / files
-
-            if (isBuffer) {
-              logger.debug(
-                `${service}.${String(method)}: received buffer as argument`,
-              )
-            }
-
-            return !isTransaction && !isBuffer
-          })
-
-          return !isTransactionOrBuffer
-        }
-
-        return !isTransaction && !isBuffer
-      })
-
+      const filteredArgs = filterArgs(args, service, method)
       try {
         const logData = {
           method: method,
