+++ libs/shared/decorators/src/lib/log-method.decorator.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { logger } from '@dmr.is/logging'
-import { Transaction } from 'sequelize'
+import { filterArgs } from './utils'
 
 export function LogMethod(logArgs: boolean | undefined = true) {
   return function (
@@ -17,30 +17,7 @@ export function LogMethod(logArgs: boolean | undefined = true) {
         category: service,
       }
 
-      const filteredArgs = args.filter((arg) => {
-        const isTransaction = arg instanceof Transaction
-        const isBuffer = Buffer.isBuffer(arg.buffer) // filter out arguments with buffer / files
-
-        if (Array.isArray(arg)) {
-          const isTransactionOrBuffer = arg.filter((a) => {
-            const isTransaction = a instanceof Transaction
-            const isBuffer = Buffer.isBuffer(a.buffer) // filter out arguments with buffer / files
-
-            if (isBuffer) {
-              logger.debug(
-                `${service}.${String(method)}: received buffer as argument`,
-              )
-            }
-
-            return !isTransaction && !isBuffer
-          })
-
-          return !isTransactionOrBuffer
-        }
-
-        return !isTransaction && !isBuffer
-      })
-
+      const filteredArgs = filterArgs(args, service, String(method))
       if (logArgs) {
         Object.assign(logData, filteredArgs)
       }
