+++ libs/shared/decorators/src/lib/utils.ts
@@ -1,6 +1,6 @@
 import { logger } from '@dmr.is/logging'
 import { ResultWrapper } from '@dmr.is/types'
-import { BaseError } from 'sequelize'
+import { BaseError, Transaction } from 'sequelize'
 import { HttpException } from '@nestjs/common'
 
 export const handleException = <T>({
@@ -94,4 +94,37 @@ export const handleException = <T>({
     code: code,
     message: message,
   })
+}
+
+/**
+ * Filters out arguments that are instances of Transaction or Buffer
+ * @param args arguments to filter
+ * @returns
+ */
+export const filterArgs = (args: any[], service?: string, method?: string) => {
+  const filteredArgs = args.filter((arg) => {
+    const isTransaction = arg instanceof Transaction
+    const isBuffer = Buffer.isBuffer(arg.buffer) // filter out arguments with buffer / files
+
+    if (Array.isArray(arg)) {
+      const isTransactionOrBuffer = arg.filter((a) => {
+        const isTransaction = a instanceof Transaction
+        const isBuffer = Buffer.isBuffer(a.buffer) // filter out arguments with buffer / files
+
+        if (isBuffer && service && method) {
+          logger.debug(
+            `${service}.${String(method)}: received buffer as argument`,
+          )
+        }
+
+        return !isTransaction && !isBuffer
+      })
+
+      return !isTransactionOrBuffer
+    }
+
+    return !isTransaction && !isBuffer
+  })
+
+  return filteredArgs
 }
