+++ libs/shared/modules/src/s3/s3.service.ts
@@ -36,7 +36,7 @@ import { IS3Service } from './s3.service.interface'
 @Injectable()
 export class S3Service implements IS3Service {
   private readonly client = new S3Client({
-    region: 'eu-west-1',
+    region: process.env.AWS_REGION ?? 'eu-west-1',
     credentials: {
       accessKeyId: process.env.AWS_ACCESS_KEY_ID ?? '',
       secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY ?? '',
